More cleanups.  Fix things so tinylogin compiles. Fix a bug in getdelim -- patch from David Whedon <dwhedon@gordian.com>.

+++ libc/unistd/sleep.c
@@ -4,7 +4,7 @@
 #include <sys/types.h>
 #include <unistd.h>
 
-void usleep(unsigned long usec)
+void usleep(unsigned int usec)
 {
 	struct timeval tv;
